requirements.txt
# Flask bits
Flask
Flask_Minify # compress webpage stuff for better serving
Flask-Login
Flask-SQLAlchemy
Flask_WTF
Bootstrap-Flask

# production-ready WSGI server
gunicorn

# fixations and visualization
numpy
opencv-python
scipy
msgpack
requests
pytest
pytest-mock
bitarray
numpy-quaternion
plotly
pandas
pathlib
matplotlib
